lan3

+++ Shopping/Models/ProductModel.cs
@@ -7,8 +7,9 @@ namespace Shopping.Models
     public class ProductModel
     {
 		[Key]
-		public int Id { get; set; }
-		[Required(ErrorMessage = "Yêu cầu nhập Tên Sản Phẩm")]
+        //public int Id { get; set; }
+        public long Id { get; set; }
+        [Required(ErrorMessage = "Yêu cầu nhập Tên Sản Phẩm")]
 		public string Name { get; set; }
 		public string Slug { get; set; }
